icon: Remove broken GRF support
The code only worked for the outdated GRF format of the initial versions
of BlocksDS. It doesn't work with modern GRF files. This means nobody can
be using this format in new projects, and fixing it is pointless because
it's a worse format than the alternatives. You can use PNG or BMP for
static icons, and GIF for animated icons. GRF only supports static icons
and it needs to be pre-converted by hand.